+++ server/mock/store.go
@@ -141,6 +141,9 @@ func (ms multiStore) SetIAVLCacheSize(size int) {
 func (ms multiStore) SetIAVLDisableFastNode(disable bool) {
 	panic("not implemented")
 }
+func (ms multiStore) SetIAVLCacheSize(size int) {
+	panic("not implemented")
+}
 
 func (ms multiStore) SetInitialVersion(version int64) error {
 	panic("not implemented")