+++ store/types/store.go
@@ -216,24 +216,6 @@ type CommitMultiStore interface {
 
 	// SetIAVLCacheSize sets the cache size of the IAVL tree.
 	SetIAVLCacheSize(size int)
-
-	// SetIAVLDisableFastNode enables/disables fastnode feature on iavl.
-	SetIAVLDisableFastNode(disable bool)
-
-	// RollbackToVersion rollback the db to specific version(height).
-	RollbackToVersion(version int64) error
-
-	// ListeningEnabled returns if listening is enabled for the KVStore belonging the provided StoreKey
-	ListeningEnabled(key StoreKey) bool
-
-	// AddListeners adds a listener for the KVStore belonging to the provided StoreKey
-	AddListeners(keys []StoreKey)
-
-	// PopStateCache returns the accumulated state change messages from the CommitMultiStore
-	PopStateCache() []*StoreKVPair
-
-	// SetMetrics sets the metrics for the KVStore
-	SetMetrics(metrics metrics.StoreMetrics)
 }
 
 //---------subsp-------------------------------